refactor: 잘못 만든 예외 제거
