--- src/main/java/com/dnd/niceteam/domain/review/exception/RecruitingNotFoundException.java
@@ -1,11 +1,0 @@
-package com.dnd.niceteam.domain.review.exception;
-
-import com.dnd.niceteam.error.exception.BusinessException;
-import com.dnd.niceteam.error.exception.ErrorCode;
-
-public class RecruitingNotFoundException extends BusinessException {
-
-    public RecruitingNotFoundException(String message) {
-        super(ErrorCode.RECRUITING_NOT_FOUND, message);
-    }
-}
